Also split out scenes integration. This is slightly inconvenient still: everything is super coupled.
